+++ esp-wifi/src/preempt_builtin/timer/xtensa.rs
@@ -1,43 +1,12 @@
-use esp_hal::interrupt::InterruptHandler;
-
+pub(crate) use crate::preempt_builtin::timer::setup_timebase as setup_timer;
 use crate::{
-    TimeBase,
-    hal::{interrupt, time::Rate, trapframe::TrapFrame, xtensa_lx, xtensa_lx_rt},
+    hal::{trapframe::TrapFrame, xtensa_lx, xtensa_lx_rt},
     preempt_builtin::task_switch,
 };
-
-/// The timer responsible for time slicing.
-const TIMESLICE_FREQUENCY: Rate = Rate::from_hz(crate::CONFIG.tick_rate_hz);
-
-use super::TIMER;
 
 // ESP32 uses Software1 (priority 3) for task switching, because it reserves
 // Software0 for the Bluetooth stack.
 const SW_INTERRUPT: u32 = if cfg!(esp32) { 1 << 29 } else { 1 << 7 };
-
-pub(crate) fn setup_timer(mut timer1: TimeBase) {
-    // The timer needs to tick at Priority 1 to prevent accidentally interrupting
-    // priority 1 limited locks.
-    timer1.set_interrupt_handler(InterruptHandler::new(
-        unsafe {
-            core::mem::transmute::<*const (), extern "C" fn()>(timer_tick_handler as *const ())
-        },
-        interrupt::Priority::Priority1,
-    ));
-    unwrap!(timer1.start(TIMESLICE_FREQUENCY.as_duration()));
-    TIMER.with(|timer| {
-        timer1.enable_interrupt(true);
-        timer.replace(timer1);
-    });
-}
-
-pub(crate) fn disable_timer() {
-    TIMER.with(|timer| {
-        let timer = unwrap!(timer.as_mut());
-        timer.enable_interrupt(false);
-        unwrap!(timer.cancel());
-    });
-}
 
 pub(crate) fn setup_multitasking() {
     unsafe {
@@ -53,25 +22,6 @@ pub(crate) fn setup_multitasking() {
 
 pub(crate) fn disable_multitasking() {
     xtensa_lx::interrupt::disable_mask(SW_INTERRUPT);
-}
-
-extern "C" fn timer_tick_handler(_context: &mut TrapFrame) {
-    TIMER.with(|timer| {
-        let timer = unwrap!(timer.as_mut());
-        timer.clear_interrupt();
-    });
-
-    // `task_switch` must be called on a single interrupt priority level only.
-    // Because on ESP32 the software interrupt is triggered at priority 3 but
-    // the timer interrupt is triggered at priority 1, we need to trigger the
-    // software interrupt manually.
-    cfg_if::cfg_if! {
-        if #[cfg(esp32)] {
-            yield_task();
-        } else {
-            task_switch(_context);
-        }
-    }
 }
 
 #[allow(non_snake_case)]