+++ esp-hal/CHANGELOG.md
@@ -67,6 +67,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - The `debug` feature has been removed. (#3425)
 - The `usb_otg` and `bluetooth` features are now considered private and have been renamed accordingly. (#3425)
 - Include `.uninit` in the `noinit` section (#3558)
+- `SoftwareInterruptControl::software_interrupt2` is no longer available when using `esp-wifi/builtin-scheduler` (#3576)
 
 ### Fixed